chore(1-3389): move strategy item container
