+++ frontend/src/component/common/StrategyItemContainer/LegacyStrategyItemContainer.tsx
