+++ frontend/src/component/common/StrategyItemContainer/StrategyItemContainer.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from '@testing-library/react';
 import { render } from 'utils/testRenderer';
-import { StrategyItemContainer } from './StrategyItemContainer';
+import { StrategyItemContainer } from './LegacyStrategyItemContainer';
 import type { IFeatureStrategy } from 'interfaces/strategy';
 
 test('should render strategy name, custom title and description', async () => {
